Moved Parameters class to seperate file.
