+++ src/CMakeLists.txt
@@ -88,7 +88,9 @@ add_library(mapping_core_base_lib
         cache/node/manager/remote_manager.cpp
         cache/node/manager/hybrid_manager.cpp
         util/gdal_source_datasets.cpp
-        datatypes/Coordinate.cpp)
+        datatypes/Coordinate.cpp
+        util/parameters.cpp
+        util/parameters.h)
 target_include_directories(mapping_core_base_lib PUBLIC ${CMAKE_CURRENT_SOURCE_DIR})
 target_link_libraries_internal(mapping_cgi mapping_core_base_lib)
 target_link_libraries_internal(mapping_manager mapping_core_base_lib)
